+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+### Changed
+- upgrade model checkpoint training to library [amrlib] 0.7.1.
+
+
 ## [0.1.0] - 2023-12-05
 ### Changed
 - Upgrade to [zensols.util] version 1.14.
@@ -28,3 +32,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [Unreleased]: https://github.com/Paul Landes/amr/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/Paul Landes/amr/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/Paul Landes/amr/compare/v0.0.0...v0.0.1
+
+[amrlib]: https://github.com/bjascob/amrlib